+++ WealthBuilder/Code/Transaction.cs
@@ -8,8 +8,8 @@ namespace WealthBuilder.Code
         public long Id { get; set; }
         public DateTime? Date { get; set; }
         public string Description { get; set; }
-        public double? Deposit { get; set; }
-        public double? Withdrawal { get; set; }
+        public decimal? Deposit { get; set; }
+        public decimal? Withdrawal { get; set; }
         public bool? Cleared { get; set; }
         public int? CheckNumber { get; set; }
         public bool? Reconciled { get; set; }
